pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "movie-shorts"
version = "0.1.0"
description = "Generate TikTok-style shorts from Real-Debrid movies."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.13.4",
  "httpx>=0.27.0",
  "pysubs2>=1.7.2",
  "python-dotenv>=1.0.1",
  "rich>=13.9.4",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]

[project.scripts]
movie-shorts = "movie_shorts.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/movie_shorts"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
